+++ src/estimation/inner_optimizer.rs
@@ -2243,6 +2243,7 @@ mod tests {
                 covariate_sigma_index: 0,
             }),
             residual_error_eta: None,
+            analytical_init: Vec::new(),
         };
 
         // Subject: 2 PK obs + 1 FREM obs
@@ -2495,6 +2496,7 @@ mod iov_tests {
             endpoints: std::collections::HashMap::new(),
             frem_config: None,
             residual_error_eta: None,
+            analytical_init: Vec::new(),
         }
     }
 
@@ -2679,6 +2681,7 @@ mod iov_tests {
             endpoints: std::collections::HashMap::new(),
             frem_config: None,
             residual_error_eta: None,
+            analytical_init: Vec::new(),
         };
         let subject = Subject {
             id: "1".into(),
@@ -2733,6 +2736,7 @@ mod iov_tests {
         let model = CompiledModel {
             frem_config: None,
             residual_error_eta: None,
+            analytical_init: Vec::new(),
             name: "noniov_mu".into(),
             has_conditional_eta_params: false,
             pk_model: PkModel::OneCptIv,